local_data_demo/requirements.txt
fastapi
uvicorn
numpy
pandas
scikit-learn
torch
torchvision
torchaudio
transformers
sentence-transformers
chromadb
beautifulsoup4
requests
python-dotenv
googlemaps
flask
flask-cors
faiss-cpu
folium
ddgs